@@ -1,27 +1,5 @@
-/*
- * @Author: Pi Patle
- * @Email: pisenliang@gmail.com
- * @Github: https://github.com/SenLiangpi
- * @Website: https://senliangpi.github.io/blog/#/
- * @Date: 2020-08-16 17:23:34
- * @LastEditors: Pi Patle
- * @LastEditTime: 2020-10-20 09:59:18
- */
-// The Vue build version to load with the `import` command
-// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-import Vue from 'vue'
-import App from './App'
+import { createApp } from 'vue'
+import App from './App.vue'
 import router from './router'
 
-import web_logging_vue from './web-logging-vue/src/index'
-Vue.use(web_logging_vue,true)
-
-Vue.config.productionTip = false
-
-/* eslint-disable no-new */
-new Vue({
-  el: '#app',
-  router,
-  components: { App },
-  template: '<App/>'
-})
+createApp(App).use(router).mount('#app')
