@@ -1,35 +1,30 @@
-<!--
- * @Author: PiPi
- * @Email: pisenliang@gmail.com
- * @Github: https://github.com/SenLiangpi
- * @Website: https://senliangpi.github.io/blog/#/
- * @Date: 2020-05-20 09:46:48
- * @LastEditors: Pi Patle
- * @LastEditTime: 2020-10-28 16:29:12
---> 
 <template>
-  <div id="app">
-    <!-- <img src="./assets/logo.png"> -->
-    <router-view/>
+  <div id="nav">
+    <router-link to="/">Home</router-link> |
+    <router-link to="/about">About</router-link>
   </div>
+  <router-view/>
 </template>
 
-<script>
-export default {
-  name: 'App',
-  created(){
-  }
-}
-</script>
-
-<style>
-* {margin: 0;padding: 0}
-/* #app {
-  font-family: 'Avenir', Helvetica, Arial, sans-serif;
+<style lang="scss">
+#app {
+  font-family: Avenir, Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
   color: #2c3e50;
-  margin-top: 60px;
-} */
+}
+
+#nav {
+  padding: 30px;
+
+  a {
+    font-weight: bold;
+    color: #2c3e50;
+
+    &.router-link-exact-active {
+      color: #42b983;
+    }
+  }
+}
 </style>
